Fixed Ace test and removed static file serving.
Relieved Server of static file-serving duty as I was advised
on SO to just use pub serve for development simplicity.. But
eventually, need to build this back in so we don't require
an external webserver just for push the client-side.

Also fixed some issues with Ace. It now properly shows up in
the browser.

Signed-off-by: Mike Lewis <mike@updroid.com>