Early startup resequenced.
The client-side classes are now created as a result of a sequence
of initialization message passing between the client and server.
Notably, how the absolute path prefix gets set in some classes
was changed.

Also fixed a bug in all the class event handlers where they were
listening on the wrong event attribute.

Signed-off-by: Mike Lewis <mike@updroid.com>