console: now compatible with interactive commands
The Console was previously incompatible with commands, programs,
and scripts that required input from the user in the middle of
execution. These would cause cmdr to crash. This is now fixed for
at least simple cases, but has not been exhaustively tested.

Note: this does not fix problems with curses-style programs, and
those will still crash cmdr.

refs #4